Fix gcc warnings due to previous commits
git-svn-id: https://svn.osgeo.org/gdal/trunk@31598 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/iso8211/ddffielddefn.cpp
@@ -583,8 +583,7 @@ char *DDFFieldDefn::ExpandFormat( const char * pszSrc )
             char       *pszContents = ExtractSubstring( pszSrc+iSrc );
             char       *pszExpandedContents = ExpandFormat( pszContents );
 
-            if( (int) (strlen(pszExpandedContents) + strlen(pszDest) + 1)
-                > nDestMax )
+            if( strlen(pszExpandedContents) + strlen(pszDest) + 1 > nDestMax )
             {
                 nDestMax = 2 * (strlen(pszExpandedContents) + strlen(pszDest));
                 pszDest = (char *) CPLRealloc(pszDest,nDestMax+1);
@@ -615,8 +614,7 @@ char *DDFFieldDefn::ExpandFormat( const char * pszSrc )
                 
             for( int i = 0; i < nRepeat; i++ )
             {
-                if( (int) (strlen(pszExpandedContents) + strlen(pszDest) + 1 + 1)
-                    > nDestMax )
+                if( strlen(pszExpandedContents) + strlen(pszDest) + 1 + 1 > nDestMax )
                 {
                     nDestMax = 
                         2 * (strlen(pszExpandedContents) + strlen(pszDest) + 1);
